# Conflicts:
#	data/__pycache__/__init__.cpython-37.pyc
#	data/__pycache__/connect_database.cpython-37.pyc
#	data/__pycache__/dataset.cpython-37.pyc
#	model_DNN/__pycache__/__init__.cpython-37.pyc
#	model_DNN/__pycache__/config.cpython-37.pyc

+++ model_cGAN/dataset.py
@@ -144,8 +144,8 @@ class TestDataSet(torch.utils.data.Dataset):
         std_set = [0.001, 0.01, 0.1, 1, 10]
         for std in std_set:
             for i in range(1000):
-                self.data_set.append(std * np.random.randn(2000))
-                self.condition_set.append(np.array([std]))
+                self.data_set.append(std * np.random.randn(2000).astype(dtype=np.float32))
+                self.condition_set.append(np.array([std], dtype=np.float32))
 
         self.data_set = np.array(self.data_set)
         self.condition_set = np.array(self.condition_set)
